format code, add README.md and last jar build

+++ src/test/groovy/org/jsalazar/fraud/service/CheckFraudUserServiceFactorySpec.groovy
@@ -4,8 +4,6 @@ import org.jsalazar.fraud.client.BankValidationService
 import org.jsalazar.fraud.client.LocationServiceClient
 import org.jsalazar.fraud.client.UserServiceClient
 import org.jsalazar.fraud.configuration.CheckFraudUserServiceConfiguration
-import org.jsalazar.fraud.service.FraudValidationServiceFactory
-import org.jsalazar.fraud.service.CheckFraudUserServiceImpl
 import spock.lang.Shared
 import spock.lang.Specification
 
@@ -42,7 +40,7 @@ class CheckFraudUserServiceFactorySpec extends Specification{
 
     def "Test create CheckFraudUserServiceImpl with factory method default"() {
         when: "calling factory method"
-        CheckFraudUserServiceImpl checkFraudUserServiceImpl = FraudValidationServiceFactory.getDefaultCheckFraudUserService(checkFraudUserServiceConfiguration)
+        CheckFraudUserServiceImpl checkFraudUserServiceImpl = FraudValidationServiceFactory.getDefaultCheckFraudUserService(checkFraudUserServiceConfiguration) as int
 
         then: "returned checkFraudUserServiceImpl should be created with methods param"
         checkFraudUserServiceImpl.userServiceClient
